+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 
 ## [Unreleased]
 
+- Add `process::{ExitCode,Termination}`.
+
 ## [0.1.14] - 2024-08-23
 
 - Add `#[must_use]` to `OwnedFd::into_raw_fd`.